pull new locale_id/placeholdr handling
